requirements.txt
flask 
flask-cors 
requests
rdflib
pyRdfa3